+++ pkg/login/keys/keytransfer_test.go
@@ -1,4 +1,4 @@
-package keytransfer
+package keys
 
 import (
 	"context"
