+++ pkg/login/login.go
@@ -6,7 +6,7 @@ import (
 	"github.com/spf13/afero"
 
 	"github.com/stripe/stripe-cli/pkg/config"
-	"github.com/stripe/stripe-cli/pkg/login/keytransfer"
+	"github.com/stripe/stripe-cli/pkg/login/keys"
 )
 
 // Login is the main entrypoint for logging in to the CLI.
@@ -16,8 +16,8 @@ func Login(ctx context.Context, baseURL string, config *config.Config) error {
 		return err
 	}
 
-	configurer := keytransfer.NewRAKConfigurer(config, afero.NewOsFs())
-	rt := keytransfer.NewRAKTransfer(configurer)
+	configurer := keys.NewRAKConfigurer(config, afero.NewOsFs())
+	rt := keys.NewRAKTransfer(configurer)
 	auth := NewAuthenticator(rt)
 	return auth.Login(ctx, links)
 }
